split utils file to fix dependency cycle
